chore(FormComponents): move under AutoExample
these are not to be exported but import-path does that
and someone might think these are public api. well they're not, these
are only for AutoExample
