+++ packages/wix-storybook-utils/src/AutoExample/components/styles.scss
