+++ packages/wix-storybook-utils/src/AutoExample/index.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {mount} from 'enzyme';
 
-import {Option} from '../FormComponents';
+import {Option} from './components';
 import AutoExample from './';
 
 class Driver {
